Move RecordingRuleViewModel
