+++ src/pages/RecordingRulesView/RecordingRulesView.tsx
@@ -3,11 +3,11 @@ import { Column, DeleteButton, EmptyState, InteractiveTable, TagList, Text, useS
 import { BackButton } from '@shared/components/Common/BackButton';
 import { HttpClientError } from '@shared/infrastructure/http/HttpClientError';
 import { getProfileMetric, ProfileMetricId } from '@shared/infrastructure/profile-metrics/getProfileMetric';
+import { RecordingRuleViewModel } from '@shared/types/RecordingRuleViewModel';
 import { PageTitle } from '@shared/ui/PageTitle';
 import React from 'react';
 
 import { EmptyLoadingPage } from '../../app/components/Onboarding/ui/EmptyLoadingPage';
-import { RecordingRuleViewModel } from './domain/RecordingRuleViewModel';
 import { useRecordingRulesView } from './domain/useRecordingRulesView';
 
 export default function RecordingRulesView() {
