+++ src/shared/types/RecordingRuleViewModel.ts
